Store user related routes in separate file
Move all user related routes like login, profile, callbacks to a new
router with prefix `/user`.
