+++ routes/main.js
@@ -2,6 +2,11 @@ const validator = require("validator");
 const markdown = require("markdown-it")();
 
 module.exports = function (app, passport) {
+	const userRoutes = require("./user")(app, passport);
+
+	// routes related to user profile and authentication
+	app.use("/user", userRoutes);
+
 	// List all courses and their scores
 	app.get("/", function (req, res) {
 		// Get a list of all courses and calculate the average scores to show
@@ -270,61 +275,6 @@ module.exports = function (app, passport) {
 			}
 		});
 	});
-
-	// logout user
-	app.get("/logout", function (req, res) {
-		req.logout();
-		res.redirect("/");
-	});
-
-	// profile page of the user
-	app.get("/profile", checkAuth, function (req, res) {
-		// Get the user's reviews
-		let sqlquery =
-			"SELECT reviews.id, \
-                        reviews.course_id, \
-                        courses.title \
-                        FROM reviews \
-                        JOIN courses \
-                        ON reviews.course_id=courses.id \
-                        JOIN users \
-                        ON reviews.user_id=users.id \
-                        WHERE reviews.user_id \
-                        LIKE ?";
-		let id = [req.user.id];
-
-		db.query(sqlquery, id, (err, result) => {
-			if (err) {
-				return console.error("Data not found: " + err.message);
-			}
-			res.render("profile.html", {
-				heading: "Profile",
-				title: "Compass - profile",
-				reviews: typeof result == "undefined" ? [] : result,
-				user: req.user,
-				infoMessage: req.flash("info"),
-				errorMessage: req.flash("error"),
-				warningMessage: req.flash("warning"),
-			});
-		});
-	});
-
-	// Intiate slack authentication process
-	app.get("/auth/slack", passport.authorize("slack.login"));
-
-	// OAuth callback url used by Slack
-	app.get(
-		"/auth/slack/callback",
-		passport.authenticate("slack.login", {
-			failureRedirect: "/",
-			failureFlash: "Slack login failed",
-		}),
-		function (req, res) {
-			// Set cookie age to 7 days
-			req.session.cookie.maxAge = 7 * 24 * 60 * 60 * 1000;
-			res.redirect("/profile");
-		}
-	);
 
 	// handle pages not found. This should be the second last route.
 	app.use(function (req, res) {
